requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
dj-database-url==0.5.0
Django==4.0.3
django-debug-toolbar==3.2.4
django-heroku==0.3.1
django-js-asset==2.0.0
django-mptt==0.13.4
django-multiselectfield==0.1.12
gunicorn==20.1.0
idna==3.3
Pillow==9.0.1
psycopg2==2.9.3
python-decouple==3.6
qrcode==7.3.1
requests==2.27.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
whitenoise==6.0.0
